0.20.20-ALPHA: - improved architecture
