+++ spring.mock.mvc/src/main/java/ru/tinkoff/qa/neptune/spring/mock/mvc/MockMvcProvider.java
@@ -1,27 +1,27 @@
-package ru.tinkoff.qa.neptune.spring.boot.starter.web.testclient;
+package ru.tinkoff.qa.neptune.spring.mock.mvc;
 
-import org.springframework.test.web.reactive.server.WebTestClient;
+import org.springframework.test.web.servlet.MockMvc;
 
 import static java.util.Objects.nonNull;
 import static org.springframework.beans.factory.BeanFactoryUtils.beanNamesForTypeIncludingAncestors;
 import static ru.tinkoff.qa.neptune.spring.boot.starter.application.contexts.CurrentApplicationContextTestExecutionListener.getCurrentApplicationContext;
 
 /**
- * Provides an instance of {@link WebTestClient}
+ * Provides an instance of {@link MockMvc}
  */
-public class WebTestClientProvider {
+final class MockMvcProvider {
 
     /**
-     * @return instance of {@link WebTestClient}
+     * @return instance of {@link MockMvc}
      */
-    public WebTestClient provide() {
+    public MockMvc provide() {
         var context = getCurrentApplicationContext();
-        var beans = beanNamesForTypeIncludingAncestors(context, WebTestClient.class);
+        var beans = beanNamesForTypeIncludingAncestors(context, MockMvc.class);
         for (var bean : beans) {
             try {
                 var result = context.getBean(bean);
                 if (nonNull(result)) {
-                    return (WebTestClient) result;
+                    return (MockMvc) result;
                 }
             } catch (Throwable e) {
                 e.printStackTrace();
