+++ spring.web.testclient/src/main/java/ru/tinkoff/qa/neptune/spring/web/testclient/WebTestClientProvider.java
@@ -1,27 +1,27 @@
-package ru.tinkoff.qa.neptune.spring.boot.starter.mock.mvc;
+package ru.tinkoff.qa.neptune.spring.web.testclient;
 
-import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.reactive.server.WebTestClient;
 
 import static java.util.Objects.nonNull;
 import static org.springframework.beans.factory.BeanFactoryUtils.beanNamesForTypeIncludingAncestors;
 import static ru.tinkoff.qa.neptune.spring.boot.starter.application.contexts.CurrentApplicationContextTestExecutionListener.getCurrentApplicationContext;
 
 /**
- * Provides an instance of {@link MockMvc}
+ * Provides an instance of {@link WebTestClient}
  */
-public final class MockMvcProvider {
+final class WebTestClientProvider {
 
     /**
-     * @return instance of {@link MockMvc}
+     * @return instance of {@link WebTestClient}
      */
-    public MockMvc provide() {
+    public WebTestClient provide() {
         var context = getCurrentApplicationContext();
-        var beans = beanNamesForTypeIncludingAncestors(context, MockMvc.class);
+        var beans = beanNamesForTypeIncludingAncestors(context, WebTestClient.class);
         for (var bean : beans) {
             try {
                 var result = context.getBean(bean);
                 if (nonNull(result)) {
-                    return (MockMvc) result;
+                    return (WebTestClient) result;
                 }
             } catch (Throwable e) {
                 e.printStackTrace();
